refactor(root): split check-ci-env's command parsing into its own module
Why
The subshell-scoping change pushed check-ci-env.ts past the repo's 500-line
cap, and the previous commit was pushed with lint errors: the verify command
was chained with `&&` after a `rg` filter, so a non-zero turbo exit never
reached the shell. Gate on the exit code, not on a grep.

What
- scripts/lib/ci-env-command.ts holds the command-level parsing — which env
  names a command assigns, the subshell scopes those assignments are in force
  for, and the repository scripts it invokes. Same pattern as
  metrics-glitter.ts being split out of metrics.ts.
- commandScopes no longer uses non-null assertions.

Verification
- bunx turbo run typecheck lint test check-ci-env
  --filter=@shepherdjerred/root-scripts --filter=@shepherdjerred/code-review
  — exit 0, 8/8 tasks
- bun scripts/check-ci-env.ts — OK, 34 scopes / 65 invocations
